# configs/openapi.yaml
openapi: "3.0.3"
info:
  title: BCK Backup Manager API
  version: "1.0.0"
  description: REST API for backup management system
  contact:
    name: BCK Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /repositories/{id}:
    get:
      tags: [Repositories]
      summary: Get repository
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        200: { description: Repository object }
    put:
      tags: [Repositories]
      summary: Update repository
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        200: { description: Updated }
    delete:
      tags: [Repositories]
      summary: Delete repository
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        204: { description: Deleted }
...